rename: 뷰어/에디터용 이미지 파일 구분
