+++ src/assets/images/Viewer/delete button.png
