+++ src/assets/images/Viewer/exit button.png
