+++ src/components/ViewEdit/ModalOptions.tsx
@@ -1,15 +1,15 @@
 import React from 'react'
 import styled from 'styled-components'
-import pdfIcon from '../../assets/images/pdf.png'
-import pdfIcon_dark from '../../assets/images/pdf_dark.svg'
-import cloudIcon from '../../assets/images/upload-cloud.png'
-import cloudIcon_dark from '../../assets/images/upload-cloud_dark.svg'
-import urlIcon from '../../assets/images/url.png'
-import urlIcon_dark from '../../assets/images/url_dark.svg'
-import qrCreateIcon from '../../assets/images/qr-code-add.png'
-import qrCreateIcon_dark from '../../assets/images/qr-code-add_dark.svg'
-import closeIcon from '../../assets/images/closeIcon.png'
-import closeIcon_dark from '../../assets/images/closeIcon_dark.svg'
+import pdfIcon from '../../assets/images/Viewer/pdf.png'
+import pdfIcon_dark from '../../assets/images/Viewer/pdf_dark.svg'
+import cloudIcon from '../../assets/images/Viewer/upload-cloud.png'
+import cloudIcon_dark from '../../assets/images/Viewer/upload-cloud_dark.svg'
+import urlIcon from '../../assets/images/Viewer/url.png'
+import urlIcon_dark from '../../assets/images/Viewer/url_dark.svg'
+import qrCreateIcon from '../../assets/images/Viewer/qr-code-add.png'
+import qrCreateIcon_dark from '../../assets/images/Viewer/qr-code-add_dark.svg'
+import closeIcon from '../../assets/images/Viewer/closeIcon.png'
+import closeIcon_dark from '../../assets/images/Viewer/closeIcon_dark.svg'
 import OptionButton from './OptionButton'
 import BackDrop from './BackDrop'
 import axios from 'axios'
