+++ src/pages/ViewerPage.tsx
@@ -1,15 +1,15 @@
 import Sidebar from '../components/ViewEdit/Sidebar/SidebarList'
 import styled from 'styled-components'
-import profile from '../assets/images/profile.png'
-import gallery from '../assets/images/gallery button.png'
-import gallery_dark from '../assets/images/gallerybutton_dark.svg'
-import version from '../assets/images/version button.png'
-import version_dark from '../assets/images/versionbutton_dark.svg'
-import exportBtn from '../assets/images/share button.png'
-import exportBtn_dark from '../assets/images/sharebutton_dark.svg'
-import deleteBtn from '../assets/images/delete button.png'
-import deleteBtn_dark from '../assets/images/deletebutton_dark.svg'
-import exit from '../assets/images/exit button.png'
+import profile from '../assets/images/Viewer/profile.png'
+import gallery from '../assets/images/Viewer/gallery button.png'
+import gallery_dark from '../assets/images/Viewer/gallerybutton_dark.svg'
+import version from '../assets/images/Viewer/version button.png'
+import version_dark from '../assets/images/Viewer/versionbutton_dark.svg'
+import exportBtn from '../assets/images/Viewer/share button.png'
+import exportBtn_dark from '../assets/images/Viewer/sharebutton_dark.svg'
+import deleteBtn from '../assets/images/Viewer/delete button.png'
+import deleteBtn_dark from '../assets/images/Viewer/deletebutton_dark.svg'
+import exit from '../assets/images/Viewer/exit button.png'
 import SidebarPanel from '../components/ViewEdit/Sidebar/SidebarPanel'
 import ModalOptions from '../components/ViewEdit/ModalOptions'
 import ModalConfirm from '../components/ViewEdit/ModalConfirm'
